requirements.txt
stripe
flake8
black
djangorestframework
django-environ
